Explicitly use IPv4 address
Starting with node.js version 17 (I believe), they switched to (effectively) preferring IPv6 (::1) when resolving `localhost`. So when vite proxies to "localhost:9000", it actually proxies to ::1:9000, and the proxying fails when nothing responds. In turn, this happens because cask interprets "localhost" as 127.0.0.1, not ::1.

So, this fix just forces both parties to use 127.0.0.1 - if the user wants IPv6, they can switch over themselves, it really does not matter for a local address, IPv4 is just easier to find answers for.